requirements.txt
aiohttp==3.10.11
httpx==0.28.1
python-dotenv==1.1.1
revolt==1.9.0
